+++ src/__screenshot_tests__/checkbox-screenshot-test.tsx
@@ -24,7 +24,7 @@ test.each(DEVICES)('Checkbox - inverse (%s)', async (device) => {
     await openStoryPage({
         id: 'components-checkbox--controlled',
         device,
-        args: {inverse: true},
+        args: {variantOutside: 'brand'},
     });
 
     const wrapper = await screen.findByTestId('checkbox-wrapper');