+++ src/__stories__/phone-number-field-lite-story.tsx
@@ -13,9 +13,17 @@ import {
 import {inspect} from 'util';
 import {phoneNumbersList} from './helpers';
 
+import type {Variant} from '../theme-variant-context';
+
 export default {
     title: 'Components/Input fields/PhoneNumberFieldLite',
     parameters: {fullScreen: true},
+    argTypes: {
+        variantOutside: {
+            options: ['default', 'brand', 'negative', 'alternative'],
+            control: {type: 'select'},
+        },
+    },
 };
 
 const getPhoneNumberSuggestions = (value: string) =>
@@ -29,7 +37,7 @@ interface PhoneNumberFieldBaseArgs {
     prefix: string;
     helperText: string;
     error: boolean;
-    inverse: boolean;
+    variantOutside: Variant;
     optional: boolean;
     showOptionalLabel: boolean;
     disabled: boolean;
@@ -43,13 +51,13 @@ const defaultBaseArgs: PhoneNumberFieldBaseArgs = {
     prefix: '',
     helperText: '',
     error: false,
-    inverse: false,
+    variantOutside: 'default',
     optional: false,
     showOptionalLabel: true,
     disabled: false,
     readOnly: false,
     preventCopy: false,
-};
+} as const;
 
 interface PhoneNumberFieldControlledArgs extends PhoneNumberFieldBaseArgs {
     initialValue: string;
@@ -85,7 +93,7 @@ const Description = () => {
 };
 
 export const Controlled: StoryComponent<PhoneNumberFieldControlledArgs> = ({
-    inverse,
+    variantOutside,
     initialValue,
     suggestions,
     ...rest
@@ -94,7 +102,7 @@ export const Controlled: StoryComponent<PhoneNumberFieldControlledArgs> = ({
     const [value, setValue] = React.useState<any>(undefined);
 
     return (
-        <ResponsiveLayout variant={inverse ? 'inverse' : undefined} fullWidth>
+        <ResponsiveLayout variant={variantOutside} fullWidth>
             <Box padding={16}>
                 <Stack space={16}>
                     <Description />
@@ -139,7 +147,7 @@ interface PhoneNumberFieldUncontrolledArgs extends PhoneNumberFieldBaseArgs {
 }
 
 export const Uncontrolled: StoryComponent<PhoneNumberFieldUncontrolledArgs> = ({
-    inverse,
+    variantOutside,
     defaultValue,
     ...rest
 }) => {
@@ -147,7 +155,7 @@ export const Uncontrolled: StoryComponent<PhoneNumberFieldUncontrolledArgs> = ({
     const [value, setValue] = React.useState<any>(undefined);
 
     return (
-        <ResponsiveLayout variant={inverse ? 'inverse' : undefined} fullWidth>
+        <ResponsiveLayout variant={variantOutside} fullWidth>
             <Box padding={16}>
                 <Stack space={16}>
                     <Description />