+++ src/__stories__/pin-field-story.tsx
@@ -2,12 +2,18 @@ import * as React from 'react';
 import {Box, Text1, Stack, ResponsiveLayout, PinField} from '..';
 import {inspect} from 'util';
 
+import type {Variant} from '../theme-variant-context';
+
 export default {
     title: 'Components/Input fields/PinField',
     parameters: {fullScreen: true},
     argTypes: {
         length: {
             control: {type: 'range', min: 1, max: 7, step: 1},
+        },
+        variantOutside: {
+            options: ['default', 'brand', 'negative', 'alternative'],
+            control: {type: 'select'},
         },
     },
 };
@@ -16,7 +22,7 @@ interface PinFieldBaseArgs {
     length: number;
     helperText: string;
     error: boolean;
-    inverse: boolean;
+    variantOutside: Variant;
     disabled: boolean;
     readOnly: boolean;
     hideCode: boolean;
@@ -26,22 +32,26 @@ const defaultBaseArgs: PinFieldBaseArgs = {
     length: 6,
     helperText: '',
     error: false,
-    inverse: false,
+    variantOutside: 'default',
     disabled: false,
     readOnly: false,
     hideCode: false,
-};
+} as const;
 
 interface PinFieldControlledArgs extends PinFieldBaseArgs {
     initialValue: string;
 }
 
-export const Controlled: StoryComponent<PinFieldControlledArgs> = ({inverse, initialValue, ...rest}) => {
+export const Controlled: StoryComponent<PinFieldControlledArgs> = ({
+    variantOutside,
+    initialValue,
+    ...rest
+}) => {
     const [rawValue, setRawValue] = React.useState<any>(initialValue);
     const [value, setValue] = React.useState<any>(undefined);
 
     return (
-        <ResponsiveLayout isInverse={inverse} fullWidth>
+        <ResponsiveLayout variant={variantOutside} fullWidth>
             <Box padding={16}>
                 <Stack space={16}>
                     <PinField
@@ -82,12 +92,16 @@ interface PinFieldUncontrolledArgs extends PinFieldBaseArgs {
     defaultValue: string;
 }
 
-export const Uncontrolled: StoryComponent<PinFieldUncontrolledArgs> = ({inverse, defaultValue, ...rest}) => {
+export const Uncontrolled: StoryComponent<PinFieldUncontrolledArgs> = ({
+    variantOutside,
+    defaultValue,
+    ...rest
+}) => {
     const [rawValue, setRawValue] = React.useState<any>(undefined);
     const [value, setValue] = React.useState<any>(undefined);
 
     return (
-        <ResponsiveLayout isInverse={inverse} fullWidth>
+        <ResponsiveLayout variant={variantOutside} fullWidth>
             <Box padding={16}>
                 <Stack space={16}>
                     <PinField