+++ src/divider.tsx
@@ -1,11 +1,11 @@
 'use client';
 import * as React from 'react';
-import {useIsInverseOrMediaVariant} from './theme-variant-context';
+import {useThemeVariant} from './theme-variant-context';
 import * as styles from './divider.css';
 
 const Divider = (): JSX.Element => {
-    const isInverse = useIsInverseOrMediaVariant();
-    return <div className={styles.variants[isInverse ? 'inverse' : 'default']} />;
+    const variant = useThemeVariant();
+    return <div className={styles.variants[variant]} />;
 };
 
 export default Divider;