+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [16.44.1](https://github.com/Telefonica/mistica-web/compare/v16.44.0...v16.44.1) (2025-12-04)
+
+
+### Bug Fixes
+
+* **Snackbar:** text color in Blau ([#1468](https://github.com/Telefonica/mistica-web/issues/1468)) ([2cd5dbe](https://github.com/Telefonica/mistica-web/commit/2cd5dbef34835bcd5838c58e81bb5f84fc30aafc))
+
+# [16.44.0](https://github.com/Telefonica/mistica-web/compare/v16.43.1...v16.44.0) (2025-12-04)
+
+
+### Features
+
+* **ThemeVariant:** deprecated inverse variant (replaced by brand) and added new negative variant ([#1466](https://github.com/Telefonica/mistica-web/issues/1466)) ([e708c22](https://github.com/Telefonica/mistica-web/commit/e708c22a4174072d402d8c6244182971bfcaaa29))
+
+## [16.43.1](https://github.com/Telefonica/mistica-web/compare/v16.43.0...v16.43.1) (2025-12-02)
+
+
+### Bug Fixes
+
+* **Chip:** close button not working with mouse/touch ([#1467](https://github.com/Telefonica/mistica-web/issues/1467)) ([87648ce](https://github.com/Telefonica/mistica-web/commit/87648ceb7df6f12cd116ae2f3b955f0f195fd286))
+
 # [16.43.0](https://github.com/Telefonica/mistica-web/compare/v16.42.2...v16.43.0) (2025-11-21)